// apps/my-mash-app-backend/src/app/interfaces/yelp.scrape.interfaces.ts
// Generated by https://quicktype.io

export interface YelpResponsed {
    businesses: YelpBusinessList[];
    total:      number;
    region?:     Region;
}

export interface YelpBusinessList {
    id:            string;
    alias:         string;
    name:          string;
    image_url:     string;
    is_closed:     boolean;
    url:           string;
    review_count:  number;
    categories:    Category[];
    rating:        number;
    coordinates:   Center;
    transactions:  string[];
    location:      Location;
    phone:         string;
    display_phone: string;
    distance:      number;
    price?:        string;
}

export interface Category {
    alias: string;
    title: string;
}

export interface Center {
    latitude:  number;
    longitude: number;
}

export interface Location {
    address1:         string;
    address2:          string;
    address3:         string;
    city:            City;
    zip_code:        string;
    country:         Country;
    state:           State;
    display_address: string[];
}

export enum City {
    Boston = "Boston",
    Brookline = "Brookline",
    Cambridge = "Cambridge",
}

export enum Country {
    Us = "US",
}

export enum State {
    Ma = "MA",
}

export interface Region {
    center: Center;
}